:tractor: Summarize all parsing in _argparse.py
