+++ deepblink/cli/_create.py
@@ -1,7 +1,6 @@
 """CLI submodule for creating a new dataset."""
 
 from typing import List, Tuple
-import argparse
 import logging
 import os
 
@@ -14,99 +13,6 @@ from ..io import basename
 from ..io import grab_files
 from ..io import load_image
 from ..util import train_valid_split
-from ._parseutil import CustomFormatter
-from ._parseutil import FolderType
-from ._parseutil import _add_utils
-
-
-def _parse_args_create(
-    subparsers: argparse._SubParsersAction, parent_parser: argparse.ArgumentParser,
-):
-    """Subparser for creation."""
-    parser = subparsers.add_parser(
-        "create",
-        parents=[parent_parser],
-        formatter_class=CustomFormatter,
-        add_help=False,
-        description=(
-            "\U0001F4BE Creation submodule \U0001F4BE\n\n"
-            "Create a custom dataset with raw files and corresponding labels. "
-            "Relies on labeling output from FIJI that was saved with the provided macro "
-            'described here "https://github.com/BBQuercus/deepBlink/wiki/Datasets".'
-        ),
-        help="\U0001F4BE Create a new dataset from raw files.",
-    )
-    group1 = parser.add_argument_group("Required")
-    group1.add_argument(
-        "-i",
-        "--input",
-        required=True,
-        type=FolderType(),
-        help=(
-            "Path to the directory containing raw images. "
-            "Note that only the specified filetypes will be processed. "
-            f"[required] [filetypes: {', '.join(EXTENSIONS)}]"
-        ),
-    )
-    group2 = parser.add_argument_group("Optional")
-    group2.add_argument(
-        "-l",
-        "--labels",
-        type=FolderType(),
-        help=(
-            "Path to the directory containing labels in csv format. "
-            "The default path accounts for using the FIJI macro described on the wiki. "
-            "[default: --INPUT/labels/]"
-        ),
-    )
-    group2.add_argument(
-        "-n",
-        "--name",
-        default="dataset",
-        type=str,
-        help=(
-            "Custom dataset name. "
-            'The file extension "npz" will be added automatically. '
-            '[default: "dataset"]'
-        ),
-    )
-    group2.add_argument(
-        "-s",
-        "--size",
-        default=None,
-        type=int,
-        help=(
-            "Image crop size. "
-            "If given, crops all images into the specified size. "
-            "Will crop non-overlapping and ignore areas that did not get covered."
-            "deepBlink requires square images powers of 2, such as 256, 512... "
-            "[default: None]"
-        ),
-    )
-    group2.add_argument(
-        "-vs",
-        "--validsplit",
-        default=0.2,
-        type=float,
-        help=(
-            "Validation split. "
-            "Split percentage (scaled between 0 - 1) of validation vs. train set. "
-            "Note the validation split is done after splitting test and trainval. "
-            "[default: 0.2]"
-        ),
-    )
-    group2.add_argument(
-        "-ts",
-        "--testsplit",
-        default=0.2,
-        type=float,
-        help=(
-            "Testing split. "
-            "Split percentage (scaled between 0 - 1) of test vs. trainval set. "
-            "[default: 0.2]"
-        ),
-    )
-    _add_utils(parser)
 
 
 class HandleCreate:
