+++ deepblink/cli/_predict.py
@@ -17,101 +17,6 @@ from ..io import load_image
 from ..io import load_model
 from ..util import delete_non_unique_columns
 from ..util import predict_shape
-from ._parseutil import CustomFormatter
-from ._parseutil import FileFolderType
-from ._parseutil import FileType
-from ._parseutil import FolderType
-from ._parseutil import ShapeType
-from ._parseutil import _add_utils
-
-
-def _parse_args_predict(
-    subparsers: argparse._SubParsersAction, parent_parser: argparse.ArgumentParser,
-):
-    """Subparser for prediction."""
-    parser = subparsers.add_parser(
-        "predict",
-        parents=[parent_parser],
-        add_help=False,
-        formatter_class=CustomFormatter,
-        description=(
-            "\U0001F914 Prediction submodule \U0001F914\n\n"
-            "Use a pre-trained model to predict blob coordinates on new data. "
-            "In addition to the required model and input file or folder, "
-            "several optional features are accessible as described below."
-        ),
-        help="\U0001F914 Predict on data with a pre-trained model.",
-    )
-    group1 = parser.add_argument_group("Required")
-    group1.add_argument(
-        "-i",
-        "--input",
-        required=True,
-        type=FileFolderType(EXTENSIONS),
-        help=(
-            "Image files to predict on. "
-            "Input can either be given as path to a directory containing files or as a single file. "
-            "The path be relative (e.g. ../dir) or absolute (e.g. /Users/myname/). "
-            "Fileglobs are currently not available. "
-            "Note that only the specified filetypes will be processed. "
-            f"[required] [filetypes: {', '.join(EXTENSIONS)}]"
-        ),
-    )
-    group1.add_argument(
-        "-m",
-        "--model",
-        required=True,
-        type=FileType(["h5"]),
-        help=(
-            "DeepBlink model. "
-            'Model has to be of file type ".h5". '
-            'The path can be relative or absolute as described in "--input". '
-            'Model can either be trained on custom data using "deepblink train" or using a pre-trained '
-            'model available through the GitHub wiki on "https://github.com/BBQuercus/deepBlink/wiki". '
-            "[required]"
-        ),
-    )
-    group2 = parser.add_argument_group("Optional")
-    group2.add_argument(
-        "-o",
-        "--output",
-        type=FolderType(),
-        help=(
-            "Output folder path. "
-            "Path to the directory into which all output files are saved. "
-            "Output files will automatically take the same name as their corresponding image. "
-            "[default: input location]"
-        ),
-    )
-    group2.add_argument(
-        "-r",
-        "--radius",
-        type=int,
-        default=None,
-        help=(
-            "Intensity radius. "
-            "If given, will calculate the integrated intensity in the specified radius around each coordinate. "
-            "If the radius is set to zero if only the central pixels intensity should be calculated. "
-            'The intensity is added as additional column to the output file called "i". '
-            "[default: None]"
-        ),
-    )
-    group2.add_argument(
-        "-s",
-        "--shape",
-        type=ShapeType(),
-        default=None,
-        help=(
-            "Image shape."
-            "Used to assess the arrangement of input image axes otherwise known as shape. "
-            "If not given, uses a basic prediction based on common defaults. "
-            'Must be in the format "(x,y,z,t,c,3)" using the specified characters. '
-            'If unsure, use "deepblink check" to determine your images shape '
-            "and more detailed information. "
-            "[default: None]"
-        ),
-    )
-    _add_utils(parser)
 
 
 class HandlePredict:
