+++ deepblink/cli/_train.py
@@ -1,59 +1,10 @@
 """CLI submodule for training."""
 
-import argparse
 import logging
 import os
 import yaml
 
 from ..training import run_experiment
-from ._parseutil import CustomFormatter
-from ._parseutil import FileType
-from ._parseutil import _add_utils
-
-
-def _parse_args_train(
-    subparsers: argparse._SubParsersAction, parent_parser: argparse.ArgumentParser,
-):
-    """Subparser for training."""
-    parser = subparsers.add_parser(
-        "train",
-        parents=[parent_parser],
-        formatter_class=CustomFormatter,
-        add_help=False,
-        description=(
-            "\U0001F686 Training submodule \U0001F686\n\n"
-            'Train a custom model using a custom dataset created in "deepblink create" '
-            "or using a published dataset."
-        ),
-        help="\U0001F686 Train a freshly baked model on a dataset.",
-    )
-    group1 = parser.add_argument_group("Required")
-    group1.add_argument(
-        "-c",
-        "--config",
-        type=FileType(["yaml"]),
-        required=True,
-        help=(
-            "Configuration file. "
-            'Path to the config.yaml created using "deepblink config". '
-            "The path be relative (e.g. ../dir) or absolute (e.g. /Users/myname/). "
-            "Please see the training information on the wiki to configure the file to your requirements. "
-            "[required]"
-        ),
-    )
-    group2 = parser.add_argument_group("Optional")
-    group2.add_argument(
-        "-g",
-        "--gpu",
-        type=int,
-        default=None,
-        help=(
-            "GPU index. "
-            "Value passed CUDA_VISIBLE_DEVICES if a GPU is used for training. "
-            "[default: None]"
-        ),
-    )
-    _add_utils(parser)
 
 
 class HandleTrain:
